+++ src/api/api_models/admission_management/CourseAdmission.ts
@@ -1,0 +1,7 @@
+export default interface CourseAdmission {
+    enrollmentId: string;
+    courseId: string;
+    moduleId: string;
+    admissionId: string;
+    timestamp: string;
+}