+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # [WIP 0.15.1](https://github.com/upb-uc4/ui-web/compare/v0.15.0...v0.15.1) (2021-XX-XX)
 ## Feature
 - add ability to fetch all information according to gdpr [#776](https://github.com/upb-uc4/ui-web/pull/776)
-
+- add course admission API [#774](https://github.com/upb-uc4/ui-web/pull/774)
 # [0.15.0](https://github.com/upb-uc4/ui-web/compare/v0.14.0...v0.15.0) (2020-12-18)
 ## Feature
 - add hyperledger versions [#754](https://github.com/upb-uc4/ui-web/pull/754)